+++ packages/core/stellar/src/print.ts
@@ -9,6 +9,13 @@ const DEFAULT_SECTION = '1. with no section';
 const STANDALONE_IMPORTS_GROUP = 'Standalone Imports';
 const MAX_USE_CLAUSE_LINE_LENGTH = 90;
 const TAB = '    ';
+export const createLevelAttributes = [`#![no_std]`];
+
+export const removeCreateLevelAttributes = (printedContract: string) =>
+  createLevelAttributes.reduce(
+    (cleanedPrintedContract, createLevelAttribute) => cleanedPrintedContract.replace(createLevelAttribute, ''),
+    printedContract,
+  );
 
 export function printContract(contract: Contract): string {
   return formatLines(
@@ -16,7 +23,9 @@ export function printContract(contract: Contract): string {
       [
         `// SPDX-License-Identifier: ${contract.license}`,
         `// Compatible with OpenZeppelin Stellar Soroban Contracts ${compatibleContractsSemver}`,
-        `#![no_std]`,
+        ...(contract.documentations.length ? ['', ...printDocumentations(contract.documentations)] : []),
+        ...(contract.securityContact ? ['', ...printSecurityTag(contract.securityContact)] : []),
+        ...createLevelAttributes,
       ],
       spaceBetween(
         printUseClauses(contract),
@@ -239,7 +248,7 @@ function printFunction(fn: ContractFunction): Lines[] {
     }
   }
 
-  return printFunction2(fn.pub, head, args, fn.tag, fn.returns, undefined, codeLines);
+  return printFunction2(fn.pub, head, args, fn.tags, fn.returns, undefined, codeLines);
 }
 
 function printContractFunctions(contract: Contract): Lines[] {
@@ -280,7 +289,7 @@ function printConstructor(contract: Contract): Lines[] {
       true,
       head,
       args.map(a => printArgument(a)),
-      undefined,
+      [],
       undefined,
       undefined,
       body,
@@ -297,15 +306,15 @@ function printFunction2(
   pub: boolean | undefined,
   kindedName: string,
   args: string[],
-  tag: string | undefined,
+  tags: string[],
   returns: string | undefined,
   returnLine: string | undefined,
   code: Lines[],
 ): Lines[] {
   const fn = [];
 
-  if (tag !== undefined) {
-    fn.push(`#[${tag}]`);
+  for (let i = 0; i < tags.length; i++) {
+    fn.push(`#[${tags[i]}]`);
   }
 
   let accum = '';
@@ -352,4 +361,12 @@ function printArgument(arg: Argument): string {
   } else {
     return `${arg.name}`;
   }
+}
+
+function printDocumentations(documentations: string[]): string[] {
+  return documentations.map(documentation => `//! ${documentation}`);
+}
+
+function printSecurityTag(securityContact: string) {
+  return ['//! # Security', '//!', `//! For security issues, please contact: ${securityContact}`];
 }